backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
dnspython
pydantic
pydantic[email]
bcrypt
python-jose[cryptography]
langchain
langchain-community
langchain-google-genai
langchain-groq
google-generativeai
google-genai
faiss-cpu
tenacity
slowapi
requests
urllib3
unstructured
pypdf
docx2txt
python-multipart
fastapi-mail
python-docx
reportlab
sentence-transformers
pytesseract
pdf2image
Pillow
pandas
pyarrow
